target: msm8916: adv7511: detect monitor connection
Detect if monitor is plugged in at boot or not, and do not enable
display if not connected.  This ensures that the simple-framebuffer
DT node is only inserted into 'chosen' if a monitor is connected.

This lets u-boot/etc fall back to serial port if there is no display
connected.